% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringostat.R
\docType{package}
\name{ringostat-package}
\alias{ringostat}
\alias{ringostat-package}
\title{ringostat: Load data from 'Ringostat API'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Loadind calls data from 'Ringostat API'. 
    See <https://help.ringostat.com/knowledge-base/article/integration-with-ringostat-via-api>.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
