% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_auth.R
\name{rs_auth}
\alias{rs_auth}
\title{Set Ringostat API key}
\usage{
rs_auth(api_key)
}
\arguments{
\item{api_key}{Your API key}
}
\value{
No return value, called for side effects
}
\description{
Set Ringostat API key
}
