% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_download_record.R
\name{rs_download_record}
\alias{rs_download_record}
\title{Download call recording}
\usage{
rs_download_record(call_id, out_file = NULL)
}
\arguments{
\item{call_id}{uniqueid of call}

\item{out_file}{name of output wav file}
}
\value{
TRUE if has been downloaded successfully
}
\description{
Download call recording
}
\examples{
\dontrun{
rs_download_record("ua-987h79879")
}
}
